+++ examples/fluid/dam_break_oil_film_2d.jl
@@ -68,7 +68,8 @@ oil_system = WeaklyCompressibleSPHSystem(oil;
                                          acceleration=(0.0, -gravity),
                                          surface_tension=SurfaceTensionAkinci(surface_tension_coefficient=0.01),
                                          correction=AkinciFreeSurfaceCorrection(oil_density),
-                                         reference_particle_spacing=fluid_particle_spacing)
+                                         reference_particle_spacing=fluid_particle_spacing,
+                                         color_value=2)
 
 # oil_system = WeaklyCompressibleSPHSystem(oil;
 #                                          smoothing_kernel, smoothing_length,