+++ src/schemes/fluid/entropically_damped_sph/system.jl
@@ -121,6 +121,7 @@ function EntropicallyDampedSPHSystem(initial_condition; smoothing_kernel, smooth
 
     surface_method = select_surface_method(surface_tension, surface_method,
                                            surface_normal_method)
+    check_surface_method_eltype(surface_method, ELTYPE)
 
     if is_colorfield_surface_method(surface_method) && reference_particle_spacing < eps()
         throw(ArgumentError("`reference_particle_spacing` must be set to a positive value when using a colorfield surface method"))